feat: utilities for repeatable response decoding (#8)
Adds a new `retch::utils` module that provides helper structs and
functions for decoding binary content.

This brings a substantial improvement over the basic `Response.text()`
method, as it implements BOM sniffing and (a variant of) byte stream
prescan. It also doesn't consume the response (byte array) in the
process.